refactor(cli): split commands + extract ChatApp hooks (phase 1)
Restructure per packages/cli/ARCHITECTURE.md phase 1. Pure moves + splits,
no behavior changes. Old paths re-export the new modules with @deprecated
JSDoc so external imports keep working through at least one minor.

- commands/{chat,run,init,doctor,config,dev,tunnel}.ts: one subcommand each
- commands/index.ts: createCli() aggregator
- commands/shared.ts: mergeWithConfig helper shared by chat + run
- app/ChatApp.tsx: view layer, pure composition
- runtime/use-runtime.ts: provider/model/apiKey/baseUrl/tools/skill state + memos
- runtime/use-tool-permissions.ts: session-wide tool allowlist + auto-approve
- runtime/use-session-meta.ts: debounced session metadata writer

All 59 existing tests pass. Type-check + build clean.
